Fix sequential e2e tests
